+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Engine | Implement language switcher on profile page
 - Engine | Add tasking abstraction
 - Engine | Performance improvement when openning questionnaires and saving QRs
+- Engine | Add debug mode to optionally load local config
 
 ### Fixed
 - ANC | Resolved a bug on measure reporting : App crash when All patient selected